Implement core workflow for local games
This commit implements the core workflow and gameplay loop for local player vs player games. This is accomplished by defining the `GameManager` trait to dictate the required behaviors of game-specific execution steps across all game types, then implementing the trait for a `LocalGameManager` (the others remain TODO).

This commit includes some other minor refactors, such as creating a separate `tui` module for terminal-based UI methods. Though, there is not much there for this at the moment.