Implement Terminal User Interface
In previous commits, the UI for FourStack was extremely rudimentary and consisted of raw terminal printing and taking unconstrained user input for column selection. This commit creates a new trait `ViewManager` to define UI methods, and implements them for `TuiManager`. This struct uses the `cursive` crate as a TUI interface with ncurses as the backend.