Refactor GameState
This commit refactors the game state fields used within the `GameManager` structs to a separate `GameState` struct, as opposed to having separate fields for the board and next player piece. This refactor makes the code a bit more modular and testable, as well as providing some forward-looking benefits for the AI and online use-cases involving parsing and transmitting game data.