Client-Server Refactor
The present architecture presents some challenges with adversarial game modes, especially with online PvP.
To combat this, I refactored out a separate `GameClient` trait with an implementation per game-mode.
This will allow for more traditional client-server interfaces which is important for online play. It is also important for
the AI gamemode, because I will need to create another commit to modify the game logic to run on a separate thread from the UI
so that the UI can display a screen while waiting for the server to return a response - for AI this means the artificial "thinking" delay,
and in online play it is waiting for the other play to take their turn. This will come in a future commit.